Improved ADC example

+++ simply/simply-iqrf-dpa-v210-examples/versions/1.0.0/src/main/java/com/microrisc/simply/iqrf/dpa/v210/examples/user_peripherals/myadc/def/MyADC.java
@@ -25,7 +25,7 @@ extends DPA_StandardServices, GenericAsyncCallable, MethodIdTransformer {
     /**
      * Gets actual value of MyADC.
      * @return actual value of MyADC <br> 
-     *         {@code Integer.MAX_VALUE} if an error has occurred during processing
+     *         {@code Float.MAX_VALUE} if an error has occurred during processing
      */
-    int get();
+    float get();
 }
